project_mercury/rtm/src/main/java/com/smartsparrow/rtm/subscription/courseware/elementthemedelete/ElementThemeDeleteRTMConsumable.java
package com.smartsparrow.rtm.subscription.courseware.elementthemedelete;

import com.smartsparrow.pubsub.data.RTMEvent;
import com.smartsparrow.rtm.message.RTMClientContext;
import com.smartsparrow.rtm.subscription.courseware.ActivityRTMSubscription;
import com.smartsparrow.rtm.subscription.courseware.message.ActivityBroadcastMessage;
import com.smartsparrow.rtm.subscription.data.AbstractRTMConsumable;

/**
 * This RTM consumable describes an element theme delete event
 */
public class ElementThemeDeleteRTMConsumable extends AbstractRTMConsumable<ActivityBroadcastMessage> {

    private static final long serialVersionUID = 2188927203639906312L;

    public ElementThemeDeleteRTMConsumable(RTMClientContext rtmClientContext, ActivityBroadcastMessage content) {
        super(rtmClientContext, content);
    }

    @Override
    public RTMClientContext getRTMClientContext() {
        return rtmClientContext;
    }

    @Override
    public String getName() {
        return String.format("author.activity/%s/%s",
                             content.getActivityId(),
                             new ElementThemeDeleteRTMEventDecoratorImpl(new ElementThemeDeleteRTMEvent()).getName(
                                     content.getElementType()));
    }

    @Override
    public String getSubscriptionName() {
        return ActivityRTMSubscription.NAME(content.getActivityId());
    }

    @Override
    public RTMEvent getRTMEvent() {
        return new ElementThemeDeleteRTMEventDecoratorImpl(new ElementThemeDeleteRTMEvent());
    }

    @Override
    public boolean equals(Object o) {
        return super.equals(o);
    }

    @Override
    public int hashCode() {
        return super.hashCode();
    }
}
